+++ android/src/main/kotlin/me/yohom/amapbase/map/marker/SimpleLoc.java
@@ -3,7 +3,7 @@ package me.yohom.amapbase.map.marker;
 public class SimpleLoc {
     public double lat;
     public double lng;
-    public float bearing;
+    public double bearing;
 
     @Override
     public String toString() {
